Move route extraction function to helper file, use it for all Router versions, prevent undefined.toLowerCase() in UrlSelect
