+++ packages/vscode-extension/lib/expo_router_plugin.js
@@ -1,10 +1,7 @@
 import { useSyncExternalStore, useEffect } from "react";
 import { useRouter } from "expo-router";
 import { store } from "expo-router/build/global-state/router-store.js";
-
-function computeRouteIdentifier(pathname, params) {
-  return pathname + JSON.stringify(params);
-}
+import { computeRouteIdentifier, extractNestedRouteList } from "./expo_router_helpers.js";
 
 function useRouterPluginMainHook({ onNavigationChange }) {
   const router = useRouter();
@@ -22,6 +19,18 @@ function useRouterPluginMainHook({ onNavigationChange }) {
 
   const displayParams = new URLSearchParams(filteredParams).toString();
   const displayName = `${pathname}${displayParams ? `?${displayParams}` : ""}`;
+
+  useEffect(() => {
+    if (!store.routeNode) {
+      return;
+    }
+    const routeList = extractNestedRouteList(store.routeNode);
+    if (global.__REACT_DEVTOOLS_GLOBAL_HOOK__?.reactDevtoolsAgent?._bridge) {
+      global.__REACT_DEVTOOLS_GLOBAL_HOOK__.reactDevtoolsAgent._bridge.send("RNIDE_routeListRetrieved", {
+        routes: routeList,
+      });
+    }
+  }, [store.routeNode]);
 
   useEffect(() => {
     onNavigationChange({
