+++ packages/vscode-extension/lib/expo_router_v5_plugin.js
@@ -1,64 +1,7 @@
 import { useEffect } from "react";
 import { useRouter } from "expo-router";
 import { store, useRouteInfo } from "expo-router/build/global-state/router-store.js";
-
-function computeRouteIdentifier(pathname, params) {
-  return pathname + JSON.stringify(params);
-}
-
-// Utility routes like _layout or +not-found should be excluded
-// routeNode.route contains the "normal" route path, which is the one that should be displayed to the user
-// However, we need to watch out for the first /index, which should be just /
-// routeNode.dynamic is null or params[], may help with letting the user fill in the dynamic parts
-// routeNode.contextKey is the full path of the route, but starts with ./ and includes parts in (brackets)
-// the URL works as is (excluding index), but those parts shouldn't be displayed in the list seen by the user
-// e.g. ./(auth)/login -> /login
-function extractNestedRouteList(rootNode) {
-  const routeList = [];
-
-  const traverse = (node) => {
-    let indexFound = false;
-    const handleIndexFound = (route) => {
-      indexFound = true;
-      return "/" + route.replace("index", "");
-    }
-
-    if (node) {
-      const { contextKey, children, route, dynamic, type } = node;
-      const fileName = route.split("/").pop();
-
-      // Won't match _layout, because for it the fileName is empty
-      // We need _layout's children, so that's necessary
-      if (fileName.startsWith("_") || fileName.startsWith("+")) {
-        return;
-      }
-      if (type === "route") {
-        routeList.push({
-          path: fileName === "index" && !indexFound ? handleIndexFound(route) : "/" + route,
-          filePath: contextKey,
-          children: children,
-          dynamic: dynamic,
-          type: type
-        });
-      }
-      if (children) {
-        children.forEach((child) => {
-          traverse(child);
-        });
-      }
-    }
-  };
-  traverse(rootNode);
-
-  return routeList.sort((a, b) => {
-    const aPath = a.path.split("/");
-    const bPath = b.path.split("/");
-    if (aPath.length === bPath.length) {
-      return a.path.localeCompare(b.path);
-    }
-    return aPath.length - bPath.length;
-  });
-}
+import { computeRouteIdentifier, extractNestedRouteList } from "./expo_router_helpers.js";
 
 function useRouterPluginMainHook({ onNavigationChange }) {
   const router = useRouter();
@@ -74,6 +17,9 @@ function useRouterPluginMainHook({ onNavigationChange }) {
   const displayName = `${pathname}${displayParams ? `?${displayParams}` : ""}`;
 
   useEffect(() => {
+    if (!store.routeNode) {
+      return;
+    }
     const routeList = extractNestedRouteList(store.routeNode);
     if (global.__REACT_DEVTOOLS_GLOBAL_HOOK__?.reactDevtoolsAgent?._bridge) {
       global.__REACT_DEVTOOLS_GLOBAL_HOOK__.reactDevtoolsAgent._bridge.send("RNIDE_routeListRetrieved", {
